requirements.txt
aiohttp==3.6.2
aiomultiprocess==0.7.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
Brotli==1.0.7
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cloudscraper==1.2.33
colorama==0.4.3
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
et-xmlfile==1.0.1
exrex==0.10.5
fire==0.3.1
geoip2==3.0.0
gevent==20.5.0
greenlet==0.4.15
idna==2.9
ipip-ipdb==1.3.2
ipython==7.14.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.0
Js2Py==0.70
loguru==0.4.1
maxminddb==1.5.2
multidict==4.7.5
openpyxl==2.4.11
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psutil==5.7.0
pycparser==2.20
Pygments==2.6.1
pyjsparser==2.7.1
PySocks==1.7.1
python-nmap==0.6.1
pytz==2019.3
records==0.5.3
requests==2.23.0
requests-file==1.4.3
requests-toolbelt==0.9.1
shodan==1.23.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
tablib==1.1.0
tenacity==6.1.0
termcolor==1.1.0
tldextract==2.2.2
tqdm==4.45.0
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.9
wcwidth==0.1.9
win32-setctime==1.0.1
XlsxWriter==1.2.8
yarl==1.4.2
